want to use this in crontab to periodically download open data that is being overwritten
The script will download datasets that the government overwrites
periodically. They do this for fear that if files are outdated, people
will analyse the data in the wrong way, which could cause headaches and
possible dangerous situations. This script was motivated by the fact
that we have  a need to preserve previous editions of datasets at the
University of Toronto Map and Data Library